tiding up the room commit

+++ Python/Ejercicios/potencias.py
@@ -1,4 +1,4 @@
-x = float(input("ingresa la base de la potencia: "))
+x = int(input("ingresa la base de la potencia: "))
 y = int(input("ingresa el exponente de la potencia: "))
 z=x**y
 print (z)
